JBEAP-672: Remove hibernate3 quickstart. Rename hibernate4 to hibernate5 as the code automatically uses the Hibernate 5 included with JBoss EAP
