--- hibernate3/src/main/java/org/jboss/as/quickstart/hibernate3/controller/MemberRegistration.java
@@ -1,110 +1,0 @@
-/*
- * JBoss, Home of Professional Open Source
- * Copyright 2015, Red Hat, Inc. and/or its affiliates, and individual
- * contributors by the @authors tag. See the copyright.txt in the
- * distribution for a full listing of individual contributors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * http://www.apache.org/licenses/LICENSE-2.0
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.jboss.as.quickstart.hibernate3.controller;
-
-import java.util.logging.Logger;
-
-import javax.annotation.PostConstruct;
-import javax.ejb.Stateful;
-import javax.enterprise.event.Event;
-import javax.enterprise.inject.Model;
-import javax.enterprise.inject.Produces;
-import javax.faces.application.FacesMessage;
-import javax.faces.context.FacesContext;
-import javax.inject.Inject;
-import javax.inject.Named;
-import javax.persistence.EntityManager;
-import org.hibernate.Session;
-
-import org.jboss.as.quickstart.hibernate3.model.Member;
-
-/**
- * @author Madhumita Sadhukhan
- */
-// The @Stateful annotation eliminates the need for manual transaction demarcation
-@Stateful
-// The @Model stereotype is a convenience mechanism to make this a request-scoped bean that has an
-// EL name
-// Read more about the @Model stereotype in this FAQ:
-// http://www.cdi-spec.org/faq/#accordion6
-@Model
-public class MemberRegistration {
-
-    @Inject
-    private Logger log;
-
-    @Inject
-    private FacesContext facesContext;
-
-    @Inject
-    private EntityManager em;
-
-    @Inject
-    private Event<Member> memberEventSrc;
-
-    private Member newMember;
-
-    @Produces
-    @Named
-    public Member getNewMember() {
-        return newMember;
-    }
-
-    @Produces
-    public void register() throws Exception {
-        log.info("Registering " + newMember.getName());
-
-        // using Hibernate session(Native API) and JPA entitymanager
-        Session session = (Session) em.getDelegate();
-        session.persist(newMember);
-
-        try {
-            memberEventSrc.fire(newMember);
-            initNewMember();
-        } catch (Exception e) {
-            // Display the reason for the error on the form
-            String errorMessage = getRootErrorMessage(e);
-            FacesMessage facesMessage = new FacesMessage(FacesMessage.SEVERITY_ERROR, errorMessage, "Registration unsuccessful");
-            facesContext.addMessage(null, facesMessage);
-        }
-    }
-
-    @PostConstruct
-    public void initNewMember() {
-        newMember = new Member();
-    }
-
-    private String getRootErrorMessage(Exception e) {
-        // Default to general error message that registration failed.
-        String errorMessage = "Registration failed. See server log for more information";
-        if (e == null) {
-            // This shouldn't happen, but return the default messages
-            return errorMessage;
-        }
-
-        // Start with the exception and recurse to find the root cause
-        Throwable t = e;
-        while (t != null) {
-            // Get the message from the Throwable class instance
-            errorMessage = t.getLocalizedMessage();
-            t = t.getCause();
-        }
-        // This is the root cause message
-        return errorMessage;
-    }
-
-}
