--- hibernate3/src/main/resources/META-INF/persistence.xml
@@ -1,38 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-    JBoss, Home of Professional Open Source
-    Copyright 2015, Red Hat, Inc. and/or its affiliates, and individual
-    contributors by the @authors tag. See the copyright.txt in the
-    distribution for a full listing of individual contributors.
-
-    Licensed under the Apache License, Version 2.0 (the "License");
-    you may not use this file except in compliance with the License.
-    You may obtain a copy of the License at
-    http://www.apache.org/licenses/LICENSE-2.0
-    Unless required by applicable law or agreed to in writing, software
-    distributed under the License is distributed on an "AS IS" BASIS,
-    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-    See the License for the specific language governing permissions and
-    limitations under the License.
--->
-<persistence xmlns="http://xmlns.jcp.org/xml/ns/persistence"
-   xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-   xsi:schemaLocation="http://xmlns.jcp.org/xml/ns/persistence http://xmlns.jcp.org/xml/ns/persistence/persistence_2_1.xsd"
-   version="2.0">
-      <persistence-unit name="primary">
-      <!-- If you are running in a production environment, add a managed 
-         data source, this example data source is just for development and testing! -->
-      <!-- The datasource is deployed as WEB-INF/hibernate3-quickstart-ds.xml, you
-         can find it in the source at src/main/webapp/WEB-INF/hibernate3-quickstart-ds.xml -->
-      <jta-data-source>java:jboss/datasources/Hibernate3QuickstartDS</jta-data-source>
-      <properties>
-         <!-- define JPA provider, check: 
-         https://docs.jboss.org/author/display/AS7/Developer+Guide#DeveloperGuide-PackagingtheHibernate3.5orgreater3.xJPApersistenceproviderwithyourapplication -->
-           <property name="jboss.as.jpa.providerModule" value="hibernate3-bundled" />
-           <!-- Define boostrat operation. This will instruct container to create tables(drop if they exist, prior to creation) -->
-         <property name="hibernate.hbm2ddl.auto" value="create-drop" />
-         <!-- Enable dump of QL/SQL queries to console -->
-         <property name="hibernate.show_sql" value="true" />
-      </properties>
-   </persistence-unit>
-</persistence>
