--- hibernate3/src/main/webapp/WEB-INF/faces-config.xml
@@ -1,41 +1,0 @@
-<?xml version="1.0"?>
-<!--
-    JBoss, Home of Professional Open Source
-    Copyright 2015, Red Hat, Inc. and/or its affiliates, and individual
-    contributors by the @authors tag. See the copyright.txt in the
-    distribution for a full listing of individual contributors.
-
-    Licensed under the Apache License, Version 2.0 (the "License");
-    you may not use this file except in compliance with the License.
-    You may obtain a copy of the License at
-    http://www.apache.org/licenses/LICENSE-2.0
-    Unless required by applicable law or agreed to in writing, software
-    distributed under the License is distributed on an "AS IS" BASIS,
-    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-    See the License for the specific language governing permissions and
-    limitations under the License.
--->
-<!-- This file is not required if you don't need any extra configuration. -->
-<faces-config version="2.2" xmlns="http://xmlns.jcp.org/xml/ns/javaee"
-   xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-   xsi:schemaLocation="
-        http://xmlns.jcp.org/xml/ns/javaee
-        http://xmlns.jcp.org/xml/ns/javaee/web-facesconfig_2_2.xsd">
-
-   <!-- This descriptor activates the JSF Servlet -->
-
-   <!-- Write your navigation rules here. You are encouraged to use CDI for 
-      creating @Named managed beans. -->
-   <validator>
-      <validator-id>validateMemberId</validator-id>
-      <validator-class>org.jboss.as.quickstart.hibernate3.util.ValidateMemberId</validator-class>
-   </validator>
-   <validator>
-      <validator-id>validateEmail</validator-id>
-      <validator-class>org.jboss.as.quickstart.hibernate3.util.ValidateEmail</validator-class>
-   </validator>
-   <validator>
-      <validator-id>validateName</validator-id>
-      <validator-class>org.jboss.as.quickstart.hibernate3.util.ValidateName</validator-class>
-   </validator>
-</faces-config>
