--- hibernate3/src/main/webapp/WEB-INF/jboss-deployment-structure.xml
@@ -1,46 +1,0 @@
-<!--
-    JBoss, Home of Professional Open Source
-    Copyright 2015, Red Hat, Inc. and/or its affiliates, and individual
-    contributors by the @authors tag. See the copyright.txt in the
-    distribution for a full listing of individual contributors.
-
-    Licensed under the Apache License, Version 2.0 (the "License");
-    you may not use this file except in compliance with the License.
-    You may obtain a copy of the License at
-    http://www.apache.org/licenses/LICENSE-2.0
-    Unless required by applicable law or agreed to in writing, software
-    distributed under the License is distributed on an "AS IS" BASIS,
-    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-    See the License for the specific language governing permissions and
-    limitations under the License.
--->
-<jboss-deployment-structure xmlns="urn:jboss:deployment-structure:1.2">
-    <deployment>
-        <exclusions>
-            <!-- Exclude container version of hibernate. By default its version 4, we want 
-                bundled version 3 loaded -->
-            <module name="org.hibernate" slot="main" />
-        </exclusions>
-        <dependencies>
-            <!-- Add dependencies, since in 'pom.xml' we exclude transitive dependencies from 
-                reactor, we need to add following modules, to make application work properly -->
-            <!-- NOTE: module name does not always correspond to maven artifact ID -->
-            <!-- Add dependency on commons.logging -->
-            <module name="org.apache.commons.logging" />
-            <!-- Add dependency on commons.collections -->
-            <module name="org.apache.commons.collections" />
-            <!-- Add dependency on log4j -->
-            <module name="org.apache.log4j" />
-            <!-- Add dependency on dom4j -->
-            <module name="org.dom4j" />
-            <!-- Add dependency on slf4j -->
-            <module name="org.slf4j" />
-            <!-- Add dependency on antlr -->
-            <module name="org.antlr" />
-            <!-- Add dependency on asm -->
-            <module name="asm.asm" />
-            <!-- Add dependency on javassist -->
-            <module name="org.javassist" />
-        </dependencies>
-    </deployment>
-</jboss-deployment-structure>
