--- hibernate3/src/main/webapp/index.xhtml
@@ -1,108 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-    JBoss, Home of Professional Open Source
-    Copyright 2015, Red Hat, Inc. and/or its affiliates, and individual
-    contributors by the @authors tag. See the copyright.txt in the
-    distribution for a full listing of individual contributors.
-
-    Licensed under the Apache License, Version 2.0 (the "License");
-    you may not use this file except in compliance with the License.
-    You may obtain a copy of the License at
-    http://www.apache.org/licenses/LICENSE-2.0
-    Unless required by applicable law or agreed to in writing, software
-    distributed under the License is distributed on an "AS IS" BASIS,
-    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-    See the License for the specific language governing permissions and
-    limitations under the License.
--->
-<ui:composition xmlns="http://www.w3.org/1999/xhtml"
-   xmlns:ui="http://java.sun.com/jsf/facelets"
-   xmlns:f="http://java.sun.com/jsf/core"
-   xmlns:h="http://java.sun.com/jsf/html"
-   template="/WEB-INF/templates/default.xhtml">
-   <ui:define name="content">
-      <h1>Welcome to JBoss!</h1>
-
-      <div>
-         <p>You have successfully deployed a Java EE web application.</p>
-      </div>
-
-      <h:form id="reg">
-         <h2>Member Registration</h2>
-         <p>Demonstrates a <b>Hibernate 3</b> application using JBoss.</p>
-         <h:panelGrid columns="3" columnClasses="titleCell">
-            <h:outputLabel for="id" value="Member Id (Enter unique number):"/>
-            <h:inputText id="id" value="#{newMember.id}" >
-               <f:validateLongRange minimum="0" maximum="999999" />
-               <f:validator validatorId="validateMemberId"/>
-             </h:inputText>            
-            <h:message for="id" errorClass="invalid"/>
-
-            <h:outputLabel for="name" value="Name:"/>
-            <h:inputText id="name" value="#{newMember.name}" >
-               <f:validateLength minimum="1" maximum="25"/>
-               <f:validator validatorId="validateName"/>
-            </h:inputText>
-            <h:message for="name" errorClass="invalid"/>
-
-            <h:outputLabel for="email" value="Email:"/>
-            <h:inputText id="email" value="#{newMember.email}">
-                <f:validateLength minimum="1"/>
-                <f:validator validatorId="validateEmail"/>
-            </h:inputText>
-            <h:message for="email" errorClass="invalid"/>
-
-            <h:outputLabel for="phoneNumber" value="Phone #:"/>
-            <h:inputText id="phoneNumber" value="#{newMember.phoneNumber}">
-               <f:validateLength minimum="9" maximum="12"/>
-               <f:validateLongRange minimum="100000000" maximum="999999999999"/>
-            </h:inputText>
-            <h:message for="phoneNumber" errorClass="invalid"/>
-
-    	    <h:outputLabel for="address" value="Address:"/>
-            <h:inputText id="address" value="#{newMember.address}" >
-               <f:validateLength minimum="1"/>
-            </h:inputText>
-            <h:message for="address" errorClass="invalid"/>
-         </h:panelGrid>
-
-         <p>
-            <h:panelGrid columns="1">
-               <h:commandButton id="register" action="#{memberRegistration.register}" value="Register" styleClass="register"/>
-            </h:panelGrid>
-           <h:panelGrid columns="1">
-               <h:messages styleClass="messages" errorClass="invalid" infoClass="valid" warnClass="warning"
-                           globalOnly="true"/>
-            </h:panelGrid>
-         </p>
-      </h:form>
-
-      <h2>Members</h2>
-      <h:panelGroup rendered="#{empty members}">
-         <em>No registered members.</em>
-      </h:panelGroup>
-      <h:dataTable var="_member" value="#{members}" rendered="#{not empty members}" styleClass="simpletablestyle">
-         <h:column>
-            <f:facet name="header">Id</f:facet>
-                #{_member.id}
-            </h:column>
-         <h:column>
-            <f:facet name="header">Name</f:facet>
-                #{_member.name}
-            </h:column>
-         <h:column>
-            <f:facet name="header">Email</f:facet>
-                #{_member.email}
-            </h:column>
-         <h:column>
-            <f:facet name="header">Phone #</f:facet>
-                #{_member.phoneNumber}
-            </h:column>
-        <h:column>
-            <f:facet name="header">Address #</f:facet>
-                #{_member.address}
-            </h:column>
-      </h:dataTable>
-   </ui:define>
-</ui:composition>
-
