+++ hibernate5/src/main/webapp/WEB-INF/beans.xml
