+++ hibernate5/src/main/webapp/resources/gfx/rhjb_eap_logo.png
